NLPtoSQL/requirements.txt
protobuf==4.25.3
python-dotenv==1.0.1
streamlit==1.33.0
